Objects&Classes
Creating a new class Factory, and then creating object instances of the class, using both primitive and non-primitive data types. The objects are then passed into an array of objects, which is then printed out by Arrays.toString(factories), requiring the default toString() method to be overriden to display the object contents properly.